[YITEAM-56] 별칭 입력 화면 구현 - JoinNameViewController
