+++ Buok/ViewControllers/User/Forget/ResetPasswordViewController.swift
@@ -5,7 +5,7 @@
 //  Created by 김혜빈 on 2021/05/09.
 //
 
-import UIKit
+import HeroUI
 
 class ResetPasswordViewController: HeroBaseViewController {
     let backButton = UIButton()
